Adding a quirk case for the MSR baremetal
It seems that our UEFI support does something weird to the usb xhci
controller.
The controller is detected as belonging to the bios and the write to the
device to transfer the ownership triggers (I think) and internal
interrupt that I could not identify.
Until I find a proper solution, I am adding a special case that skips
the "nice" ownership transfer and does what linux would do with
a misbehaving bios, i.e., steal the ownership of the device.
This seems to work right now on the hardware, but we need to revisit
this.