Adding support for finer grained traps
We want to disambiguate between various external interrupts to
selectively trap them.

Now the monitor is responsible for guaranteeing the uniformity of filter
interrupts. Specifically, we do not want the exception bitmap and the
external interrupt exiting to be explicitly set by a client, but rather
automatically derived from the domain's configuration and "frozen" once
we start allocating core contexts.