Saving the instrumentation for the PIT bug.
We have a guest that starts slow and gets stuck in the calibrate loop in
the boot of linux. This is because PIT timer arrives when LVT0 is masked
by the guest so it is not delivered.
PIT emulation after that does not bother attempting to reinject when the
TIMER fires again, while the guest is stuck in a loop waiting for
jiffies.