Patch to avoid the LVT0 mask bug in KVM
Our td1 apparently boot slightly slower and receives an interrupt while
LVT0 is masked. This drops the PIT timer. Meanwhile, the PIT emulation
thinks there is a pending timer and refuses to reinject or re-notify the
vcpu. We fix this corner case by asking a KVM_REQ_EVENT on the PIC core
when running into a pending event.